+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # V1.3
+- Alerts can now have a glow set.
+- Elements that require Preview Mode, have them.
+- Added `Innervate Alert` to Quality of Life → Alerts.
 - Added `Health: Over Absorbs` to ElvUI Enhancements, using ElvUI's absorb colour and texture.
 - Tweak `KeystoneReroll` a bit.
 - Add `Dungeon casts` Module. This will track all nearby casts from enemies.